refactor: lv1 패키지 비활성화 및  lv2 패키지로 코드 이동
- lv1 패키지의 모든 내용 주석 처리하여 비활성화
- lv2 브랜치에서 lv2 패키지에 lv1 패키지 내용 copy&paste
